data_split_automation/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.0
openpyxl>=3.0.0  # For Excel file support 